+++ modules/contracts/src.sol/lib/LibUtils.sol
@@ -2,6 +2,10 @@
 pragma solidity ^0.7.1;
 pragma experimental ABIEncoderV2;
 
+/// @title LibUtils
+/// @author Connext <support@connext.network>
+/// @notice Contains a helper to revert if a call was not successfully
+///         made
 library LibUtils {
     // If success is false, reverts and passes on the revert string.
     function revertIfCallFailed(bool success, bytes memory returnData)