+++ modules/contracts/src.sol/transferDefinitions/TransferDefinition.sol
@@ -5,6 +5,12 @@ pragma experimental ABIEncoderV2;
 import "../interfaces/ITransferDefinition.sol";
 import "../interfaces/ITransferRegistry.sol";
 
+/// @title TransferDefinition
+/// @author Connext <support@connext.network>
+/// @notice This contract helps reduce boilerplate needed when creating
+///         new transfer definitions by providing an implementation of
+///         the required getter
+
 abstract contract TransferDefinition is ITransferDefinition {
     function getRegistryInformation()
         external