+++ modules/contracts/src.sol/transferDefinitions/Withdraw.sol
@@ -6,6 +6,7 @@ import "./TransferDefinition.sol";
 import "../lib/LibChannelCrypto.sol";
 
 /// @title Withdraw
+/// @author Connext <support@connext.network>
 /// @notice This contract burns the initiator's funds if a mutually signed
 ///         withdraw commitment can be generated
 
@@ -27,6 +28,7 @@ contract Withdraw is TransferDefinition {
         bytes responderSignature;
     }
 
+    // Provide registry information
     string public constant override Name = "Withdraw";
     string public constant override StateEncoding =
         "tuple(bytes initiatorSignature, address initiator, address responder, bytes32 data, uint256 nonce, uint256 fee, address callTo, bytes callData)";
@@ -39,8 +41,10 @@ contract Withdraw is TransferDefinition {
         override
         returns (bool)
     {
+        // Get unencoded information
         TransferState memory state = abi.decode(encodedState, (TransferState));
         Balance memory balance = abi.decode(encodedBalance, (Balance));
+
         require(balance.amount[1] == 0, "Withdraw: NONZERO_RECIPIENT_BALANCE");
         require(
             state.initiator != address(0) && state.responder != address(0),
@@ -59,6 +63,8 @@ contract Withdraw is TransferDefinition {
             ),
             "Withdraw: INVALID_INITIATOR_SIG"
         );
+        
+        // Valid initial transfer state
         return true;
     }
 
@@ -72,8 +78,9 @@ contract Withdraw is TransferDefinition {
             abi.decode(encodedResolver, (TransferResolver));
         Balance memory balance = abi.decode(encodedBalance, (Balance));
 
-        // Allow for a withdrawal to be canceled if an empty signature is passed in
-        // Should have *specific* cancellation action, not just any invalid sig
+        // Allow for a withdrawal to be canceled if an empty signature is 
+        // passed in. Should have *specific* cancellation action, not just
+        // any invalid sig
         bytes memory b = new bytes(65);
         if (keccak256(resolver.responderSignature) == keccak256(b)) {
             // Withdraw should be cancelled, no state manipulation needed