+++ modules/protocol/src/testing/vector.spec.ts
@@ -78,7 +78,7 @@ describe("Vector", () => {
 
   describe("Vector.setup", () => {
     let vector: Vector;
-    const counterpartyIdentifier = "indra6LkSoBv6QD5BKZ5vZQnVsd8cq6Tyb2oi93s62sTvW6xUUQg8PC";
+    const counterpartyIdentifier = getRandomChannelSigner().publicIdentifier;
 
     beforeEach(async () => {
       const signer = getRandomChannelSigner();
@@ -114,7 +114,7 @@ describe("Vector", () => {
         transferRegistryAddress: mkAddress("0xdef"),
       };
       const validParams = {
-        counterpartyIdentifier: mkPublicIdentifier(),
+        counterpartyIdentifier,
         networkContext: { ...network },
         timeout: "1000",
       };
@@ -128,7 +128,7 @@ describe("Vector", () => {
         {
           name: "should fail if there is an invalid counterparty",
           params: { ...validParams, counterpartyIdentifier: "fail" },
-          error: 'should match pattern "^indra([a-zA-Z0-9]{50})$"',
+          error: 'should match pattern "^vector([a-zA-Z0-9]{50})$"',
         },
         {
           name: "should fail if there is no transferRegistryAddress",